fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
motor==3.3.2
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9